+++ idearium-lib/common/kue.js
@@ -2,6 +2,12 @@
 
 const log = require('./log')('idearium-lib:common:kue');
 const kueQueue = require('./kue-queue');
+const path = require('path');
+const { Loader } = require('../');
+
+// Setup the loader.
+const loader = new Loader();
+loader.camelCase = false;
 
 /**
  * Verify and create a factory method for creating kue jobs.
@@ -61,19 +67,48 @@ const constructCreateMethod = (type, properties) => {
 };
 
 /**
- * Helper function to process Kue jobs.
- * @param {String} type Job type to process.
- * @param {Function} fn Queue process callback.
- * @return {Void} Processes the job.
+ * Load jobs from the jobs directory.
+ * @return {Promise} Resolves with the jobs.
  */
-const process = (type, fn) => kueQueue.process(type, (job, done) => {
+const loadJobs = () => loader.load(path.join(process.cwd(), 'jobs'));
 
-    log.debug({ job, type }, `Processing job with type ${type}`);
+/**
+ * Process all jobs in the jobs directory.
+ * @return {Void} Processes the jobs.
+ */
+const processJobs = () => new Promise((resolve, reject) => {
 
-    fn(job)
-        .then(data => done(null, data))
-        .catch(done);
+    loadJobs()
+        .then((jobs) => {
+
+            const jobKeys = Object.keys(jobs);
+
+            // An array of job process functions.
+            const jobProcesses = jobKeys.map((type) => new Promise((resolve, reject) => {
+
+                kueQueue.process(type, (job, done) => {
+
+                    log.debug({ job, type }, `Processing job with type ${type}`);
+
+                    const fn = jobs[type];
+
+                    log.debug({ fn }, 'Calling function');
+
+                    fn(job)
+                        .then(data => resolve(done(null, data)))
+                        .catch(err => reject(done(err)));
+
+                });
+
+            }));
+
+            // Process the jobs.
+            return Promise.all(jobProcesses);
+
+        })
+        .then(resolve)
+        .catch(reject);
 
 });
 
-module.exports = { constructCreateMethod, process };
+module.exports = { constructCreateMethod, processJobs };