+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is a history of the changes made to idearium-lib.
 
+## Unreleased
+
+- Add `MandrillEmail` to common.
+- Replace `kue.process` with `kue.processJobs`. This now automates processing all jobs in the `jobs` directory.
+
 ## 1.0.0-alpha.14
 
 - Simplified `common/exception` now that Opbeat should become standard practice.